requirements.txt
boto3==1.12.25
pyyaml=5.3.1
redis==3.4.1
ecdsa==0.15
pycryptodome==3.9.7